backend/requirements.txt
Flask
python-dotenv
psycopg2-binary
PyJWT
flask-cors
requests
python-dateutil
